svc_b/requirements.txt
flask==1.1.2
azure-storage-blob==12.3.2
apscheduler==3.6.3